refactor(advisory): merge CSAF tabs into single Tabs component
Eliminates dual useTabControls hooks sharing the same URL param key
`ad:activeTab` by integrating CSAF-specific tabs directly into
advisory-details.tsx with conditional rendering based on advisory type.

Removes the separate CsafAdvisoryTabs component.

Implements TC-4670
